requirements.txt.txt
spotipy
yt-dlp
mutagen
tqdm
python-dotenv
